+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Version 0.0.x
 
+### 0.0.3.x (2020-Dec-22)
+
+- Added example of analogInput.reliability, and binaryInput.reliability
+- Added example of device.vendorname, and device.vendoridentifier
+- Added bacnet_reliability enumeration to CASBACnetStackAdapter.py
+
 ### 0.0.2.x (2020-Dec-11)
 
 - Added units